Add missing eUSCI regs, CTLW0 fields
Merge CTL0 and CTL1 into CTLW0.
Merge BR0 and BR1 into BRW.
Add CTLW0 fields, RXBUF, TXBUF.
Fix register widths (e.g. STATW).
Merge UCOA bits

+++ src/usci_a0_spi_mode/uca0ie_spi.rs
@@ -37,7 +37,7 @@ impl W {
 #[doc = "USCI A0 Interrupt Enable Register\n\nYou can [`read`](crate::Reg::read) this register and get [`uca0ie_spi::R`](R). You can [`reset`](crate::Reg::reset), [`write`](crate::Reg::write), [`write_with_zero`](crate::Reg::write_with_zero) this register using [`uca0ie_spi::W`](W). You can also [`modify`](crate::Reg::modify) this register. See [API](https://docs.rs/svd2rust/#read--modify--write-api)."]
 pub struct Uca0ieSpiSpec;
 impl crate::RegisterSpec for Uca0ieSpiSpec {
-    type Ux = u8;
+    type Ux = u16;
 }
 #[doc = "`read()` method returns [`uca0ie_spi::R`](R) reader structure"]
 impl crate::Readable for Uca0ieSpiSpec {}
