+++ src/usci_a0_spi_mode/uca0statw_spi.rs
@@ -65,7 +65,7 @@ impl W {
 #[doc = "USCI A0 Status Register\n\nYou can [`read`](crate::Reg::read) this register and get [`uca0statw_spi::R`](R). You can [`reset`](crate::Reg::reset), [`write`](crate::Reg::write), [`write_with_zero`](crate::Reg::write_with_zero) this register using [`uca0statw_spi::W`](W). You can also [`modify`](crate::Reg::modify) this register. See [API](https://docs.rs/svd2rust/#read--modify--write-api)."]
 pub struct Uca0statwSpiSpec;
 impl crate::RegisterSpec for Uca0statwSpiSpec {
-    type Ux = u8;
+    type Ux = u16;
 }
 #[doc = "`read()` method returns [`uca0statw_spi::R`](R) reader structure"]
 impl crate::Readable for Uca0statwSpiSpec {}
