+++ src/usci_a1_spi_mode/uca1ie_spi.rs
@@ -37,7 +37,7 @@ impl W {
 #[doc = "USCI A1 Interrupt Enable Register\n\nYou can [`read`](crate::Reg::read) this register and get [`uca1ie_spi::R`](R). You can [`reset`](crate::Reg::reset), [`write`](crate::Reg::write), [`write_with_zero`](crate::Reg::write_with_zero) this register using [`uca1ie_spi::W`](W). You can also [`modify`](crate::Reg::modify) this register. See [API](https://docs.rs/svd2rust/#read--modify--write-api)."]
 pub struct Uca1ieSpiSpec;
 impl crate::RegisterSpec for Uca1ieSpiSpec {
-    type Ux = u8;
+    type Ux = u16;
 }
 #[doc = "`read()` method returns [`uca1ie_spi::R`](R) reader structure"]
 impl crate::Readable for Uca1ieSpiSpec {}
